fixed add method

+++ src/main/java/com/lawson/chavoris/calculator/abstractions/BaseMathCalculator.java
@@ -12,19 +12,19 @@ public abstract class BaseMathCalculator {
     public abstract void total();
 
     public abstract void help();
-    public abstract void add(int a,int b);
-    public abstract void add(int a);
+    public abstract void add(double a,double b);
+    public abstract void add(double a);
 
-    public abstract void subtract(int a, int b);
-    public abstract void subtract(int a);
+    public abstract void subtract(double a, double b);
+    public abstract void subtract(double a);
 
-    public abstract void multiply(int a, int b);
-    public abstract void multiply(int a);
+    public abstract void multiply(double a, double b);
+    public abstract void multiply(double a);
 
-    public abstract void divide(int a, int b);
-    public abstract void divide(int a);
+    public abstract void divide(double a, double b);
+    public abstract void divide(double a);
 
     public abstract void powOf2();
-    public abstract void powOf2(int a);
+    public abstract void powOf2(double a);
 
 }
